backend/requirements.txt
asgiref==3.11.0
Django==6.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.4.0
django-cors-headers==4.6.0
drf-spectacular==0.28.0
dotenv==0.9.9
python-dotenv==1.2.1
sqlparse==0.5.5
PyJWT==2.7.0
